% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer_fit.R
\name{add_forward_backward}
\alias{add_forward_backward}
\title{Determine forward/backward curve}
\usage{
add_forward_backward(wafer)
}
\arguments{
\item{wafer}{Data from a single wafer}
}
\description{
Adds a new column called `direction` indicating if the row relates
to forward or backward.
}

