% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_curves.R
\name{calc_curves}
\alias{calc_curves}
\title{Calculate curves}
\usage{
calc_curves(fitted, param)
}
\arguments{
\item{fitted}{Fitted data.frame as given by \code{\link{fit_wafer}}.}

\item{param}{Curve parameters as given by \code{\link{sample_under_param}} or \code{\link{week_param}}}
}
\value{
A list containing elements \code{forward} and \code{backward}, both
containing one curve per row corresponding to the curves for each week provided.
}
\description{
Calculates both forwards and backwards curves using curve parameters
as given by \code{\link{sample_under_param}} or \code{\link{week_param}}.
}
\examples{
wafers_folder = file.path(path.package("voltagefit"),"extdata") # path to wafers data directory
fitted = fit_all(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), replicate = 1:6, treatment = rep(1,6))
fitman = fit_manova(fitted, design)
underp = under_param(fitman)
param = sample_under_param(underp)
calc_curves(fitted,param)

}

