% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_models.R
\name{curve_5BARO_derivs}
\alias{curve_5BARO}
\alias{curve_5BARO_derivs}
\title{5BARO model voltage curve}
\usage{
curve_5BARO_derivs(param, x)

curve_5BARO(x, param)
}
\arguments{
\item{param}{list consisting of 5 parameters for the curve function:
\describe{
     \item{theta_1}{Value of y at the curve minimum}
     \item{theta_2}{The range of response. Note: (theta_1+theta_2) gives 
     the value of y at the curve minimum}
     \item{theta_3}{A measure of the steepness of the curve}
     \item{theta_4}{Value of x shift applied to the data}
     \item{theta_5}{A measure of the asymmetry}
  }}

\item{x}{a range of values for which the model curve is evaluated}
}
\value{
A range of values, corresponding to the value of the 5BARO curve for 
the given parameters, over the provided range of x values.
}
\description{
Model voltage curve based on the 5BARO model. The first derivatives with 
respect to each parameter can also be calculated.
}

