% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_models.R
\name{curve_5PL_derivs}
\alias{curve_5PL}
\alias{curve_5PL_derivs}
\title{5PL model voltage curve}
\usage{
curve_5PL_derivs(param, x)

curve_5PL(x, param)
}
\arguments{
\item{param}{list consisting of 5 parameters for the curve function:
\describe{
     \item{theta_1}{Value of y at the curve minimum}
     \item{theta_2}{Value of y at the curve maximum}
     \item{theta_3}{Value of x (+100) at the point of inflection}
     \item{theta_4}{A measure of the steepness of the curve}
     \item{theta_5}{A measure of the asymmetry}
  }}

\item{x}{a range of values for which the model curve is evaluated}
}
\value{
A range of values, corresponding to the value of the 5PL curve for 
the given parameters, over the provided range of x values.
}
\description{
Model voltage curve based on the 5PL model. The first derivatives with 
respect to each parameter can also be calculated. 
NOTE: Here the data is shifted >> 0 in x, as the raw curve is not designed 
to handle x values of less than zero.
}

