% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_diff.R
\name{curve_diff_mean}
\alias{curve_diff_mean}
\alias{curve_diff_sample}
\title{Curve differences}
\usage{
curve_diff_mean(x, type, direction = "Forward")

curve_diff_sample(x, type, n, direction = "Forward")
}
\arguments{
\item{x}{Data.frame as output by \code{\link{fit_manova}}.}

\item{type}{A vector of length two containing the comparision}

\item{direction}{Forward or Backward.}

\item{n}{a positive integer; the number of samples to generate.}
}
\description{
Generate difference between two treatment/week effects.
}

