% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_models.R
\name{curve_exp_poly}
\alias{curve_4BARO}
\alias{curve_4PL}
\alias{curve_5BARO}
\alias{curve_5PL}
\alias{curve_exp_poly}
\alias{curve_tanh}
\title{Model device voltage curves}
\usage{
curve_exp_poly(x, param)

curve_tanh(x, param)

curve_4PL(x, param)

curve_5PL(x, param)

curve_4BARO(x, param)

curve_5BARO(x, param)
}
\arguments{
\item{x}{a range of values for which the model curve is evaluated}

\item{param}{list of parameters for the model curve function}
}
\value{
A range of values, corresponding to the value of the model curve for 
the given parameters, over the provided range of x values.
}
\description{
Functions implementing voltage curves. Several curve types are available 
for use. Each type of curve can be varied through a list of parameters.
}

