% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_models.R
\name{curve_piecewise}
\alias{curve_piecewise}
\title{Piecewise power model voltage curve}
\usage{
curve_piecewise(x, param)
}
\arguments{
\item{x}{a range of values for which the model curve is evaluated}

\item{param}{list consisting of 8 parameters for the curve function:}
}
\value{
A range of values, corresponding to the value of the curve for 
the given parameters, over the provided range of x values.
}
\description{
Model voltage curve based on a piecewise constant, quadratic then power fit.
}
