% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_models.R
\name{curve_power}
\alias{curve_power}
\title{Piecewise power model voltage curve}
\usage{
curve_power(x, param)
}
\arguments{
\item{x}{a range of values for which the model curve is evaluated}

\item{param}{list consisting of 4 parameters for the curve function:}
}
\value{
A range of values, corresponding to the value of the curve for 
the given parameters, over the provided range of x values.
}
\description{
Model voltage curve based on a power fit. Two pmax are used. pmax(0,x) 
removes the -ve x part of the curve. pmax(log(1e-13),v) is used to cap 
the minimum value to log(1e-13).
}
