% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_models.R
\name{curve_tanh}
\alias{curve_tanh}
\title{Tanh model voltage curve}
\usage{
curve_tanh(x, param)
}
\arguments{
\item{x}{a range of values for which the model curve is evaluated}

\item{param}{list consisting of 4 parameters for the curve function:
\describe{
     \item{theta_1}{Value of y at center of the curve}
     \item{theta_2}{Distance from the center of the curve to the top/bottom}
     \item{theta_3}{A measure of the steepness of the curve}
     \item{theta_4}{Value of x shift applied to the data}
  }}
}
\value{
A range of values, corresponding to the value of the tanh curve for 
the given parameters, over the provided range of x values.
}
\description{
Model voltage curve based on a tanh sigmoidal shape. The first derivatives with 
respect to each parameter can also be calculated.
}

