% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{curves}
\alias{curves}
\title{Function to create curves}
\usage{
curves(fitted, param)
}
\arguments{
\item{fitted}{Fitted data.frame as given by \code{\link{fitwafer}}.}

\item{param}{Samples as given by \code{\link{undercurvesim}} or \code{\link{weekparam}}.}
}
\value{
A list containing elements forward and backward, each containing the curves for each week (one curve per row).
}
\description{
Function to create curves
}
\examples{
wafers_folder = paste(path.package("voltagefit"),"/extdata/",sep="") # path to wafters data directory
fitted = fitall(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), replicate = c(1:6), treatment = rep(1:6))
fitman = fitmanova(fitted, design)
underp = underparam(fitman)
param = undercurvesim(underp)
curves(fitted,param)

}

