% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer_fit.R
\name{fit_all}
\alias{fit_all}
\title{Fit curves to multiple wafers}
\usage{
fit_all(path, maxit = 10000, verbose = TRUE)
}
\arguments{
\item{path}{Directory containing multiple .rds files, each containing a single wafer}

\item{maxit}{(Optional) Maxiumum number of iterations for use in optimiser (Default: 10000)}

\item{verbose}{(Optional) Print verbose output (Default: \code{TRUE})}
}
\value{
A data.frame consisting of the fields:
  \describe{
     \item{id}{Wafer ID}
     \item{name}{Each device on the wafer}
     \item{max}{Maximum of \code{log(abs(ID))}, where \code{ID} is the current drain data taken from \code{wafer}}
     \item{cost}{The value of the cost function for each device}
     \item{direction}{Whether the curve direction is forward or backward}
     \item{X1 ... X6}{The parameters characterising the curve}
  }
  The attribute v (voltage gate readings for one device), used in curve
  functions and plotting functions, is also appended.
}
\description{
Fits curves to multiple wafers in a directory, using \code{\link{fit_wafer}}, 
returning the parameters and cost for both forward and backwards curves.
}
\examples{
wafers_folder = file.path(path.package("voltagefit"),"extdata") # path to wafers data directory
fit = fit_all(wafers_folder)

}

