% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manova.R
\name{fit_manova}
\alias{fit_manova}
\title{Perform MANOVA analysis using fitted curves}
\usage{
fit_manova(fitted, design)
}
\arguments{
\item{fitted}{Fitted data.frame as given by \code{\link{fit_wafer}}.}

\item{design}{The design matrix (of type data.frame with columns week, wafer, replicate, treatment).}
}
\value{
A list with elements forward and backward, each containing
the MANOVA output incorporating weights and vcov (excluding weights).
}
\description{
Performs a MANOVA analysis using curves that have been fitted
to wafers, using \code{\link{fit_wafer}}.
}
\examples{
wafers_folder = file.path(path.package("voltagefit"),"extdata") # path to wafers data directory
fitted = fit_all(wafers_folder,dev_curve=curve_5BARO)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), 
              treatment = rep(1,6))
fitman = fit_manova(fitted, design)

}

