% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer_fit.R
\name{trans_device}
\alias{fit_wafer}
\alias{trans_device}
\alias{validate_device}
\title{Fit logcurves to a wafer}
\usage{
trans_device(wafer)

validate_device(wafer)

fit_wafer(wafer, trans = trans_device, validate = validate_device,
  initparams = c(-10, -30, 5, 0, 5), maxit = 10000, verbose = TRUE)
}
\arguments{
\item{wafer}{Data from a single wafer.}

\item{trans}{Wafer ransform function}

\item{validate}{Wafer validate function}

\item{initparams}{Inital parameter guess for use in optimiser}

\item{maxit}{Maxiumum number of iterations for use in optimiser (Default: 10000).}

\item{verbose}{Print verbose output (Default: \code{TRUE}).}
}
\value{
A data.frame consisting of the fields:
  \describe{
     \item{id}{Wafer ID}
     \item{name}{Each device on the wafer}
     \item{cost}{The value of the cost function for each device}
     \item{direction}{Whether the curve direction is forward or backward}
     \item{X1 ... X6}{The parameters characterising the curve}
  }
}
\description{
Fits logcurves to a wafer and returns the curve parameters and resulting
cost for both forward and backwards curves. It is assumed that any wafer
entering this function has been pre-filtered. See the function "filtered"
for details of how wafers are filtered.
}
\examples{
fit_wafer(wafer3737)

}

