% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer_fit.R
\name{fit_wafer_and_plot}
\alias{fit_wafer_and_plot}
\title{fit_wafer_and_plot
Fits given logcurves a wafer and plots the devices along with each fit.}
\usage{
fit_wafer_and_plot(wafer, logc, minlogc, initpars, maxit = 10000)
}
\arguments{
\item{wafer}{Data from a single wafer.}

\item{logc}{logcurve function}

\item{minlogc}{minlogcurve function}

\item{initpars}{inital parameters}

\item{maxit}{Maxiumum number of iterations for use in optimiser (Default: 10000).}
}
\description{
fit_wafer_and_plot
Fits given logcurves a wafer and plots the devices along with each fit.
}

