% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer_fit.R
\name{fit_wafer_with_curve_diff}
\alias{fit_wafer_with_curve_diff}
\title{fit_wafer_with_curve_diff}
\usage{
fit_wafer_with_curve_diff(wafer, logc, minlogc, diffminlogc, initpars,
  maxit = 10000)
}
\arguments{
\item{wafer}{Data from a single wafer.}

\item{logc}{logcurve function}

\item{minlogc}{minlogcurve function}

\item{initpars}{inital parameters}

\item{maxit}{Maxiumum number of iterations for use in optimiser (Default: 10000).}
}
\description{
fit_wafer_with_curve_diff
}

