% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fitall}
\alias{fitall}
\title{Fit logcurve to multiple wafers}
\usage{
fitall(path, iterlim = 2000)
}
\arguments{
\item{path}{Directory where multiple .rds files, each containing a single wafer, are located}

\item{iterlim}{Maxiumum number of iterations to use in nlm (default=2000).}
}
\value{
A data.frame consisting of the fields:\cr
  id - wafer id\cr
  name - each device on the wafer\cr
  max - maximum of log(abs(wafer$ID))\cr
  ID = current drain\cr
  cost - the value of costfun for each device\cr
  direction - whether it is the forward or backward curve\cr
  X1 ... X6 - the parameters characterising the curves\cr\cr
  The attribute v (voltage gate readings for one device), used in curve functions and plotting functions, is appended.
}
\description{
Fits logcurves to multiple wafers, given in a directory, and returns the parameters and cost 
for both forward and backwards curves. An extended version of fitwafer. It is assumed that 
any wafer entering this function (and subsequently fitwafer) has been pre-filtered. 
See the function "filtered" for details of how wafers are filtered.
}
\examples{
wafers_folder = paste(path.package("voltagefit"),"/extdata/",sep="") # path to wafters data directory
fit = fitall(wafers_folder)

}

