% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fitwafer}
\alias{fitwafer}
\title{Fit logcurve to a wafer}
\usage{
fitwafer(wafer, iterlim = 2000, id = NULL)
}
\arguments{
\item{wafer}{Data from a single wafer.}

\item{iterlim}{Maxiumum number of iterations to use in nlm (Default: 2000).}

\item{id}{WaferID (Default: A random string is generated as the ID).}
}
\value{
A data.frame consisting of the fields:\cr
  id - wafer id\cr
  name - each device on the wafer\cr
  max - maximum of log(abs(wafer$ID))\cr
  ID = current drain\cr
  cost - the value of costfun for each device\cr
  direction - whether it is the forward or backward curve\cr
  X1 ... X6 - the parameters characterising the curves\cr
}
\description{
Fits logcurves to a wafer and returns the parameters and cost for both forward and backwards curves. 
It is assumed that any wafer entering this function has been pre-filtered. 
See the function "filtered" for details of how wafers are filtered.
}
\examples{
fitwafer(wafer3737)

}

