% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Calculate curve parameters for each week}
\usage{
get_params(fm, combine = FALSE)
}
\arguments{
\item{fm}{Fitted manova list, as output by \code{\link{fit_manova}}.}
}
\value{
A data.frame consisting of the fields:
  \describe{
     \item{direction}{Whether the curve direction is forward or backward}
     \item{week}{The week to which the curve corresponds}
     \item{X1 ... X6}{The parameters characterising the week curve}
  }
}
\description{
Calculate the parameters characterising both the forward and backward curves for
each week present in a fitted manova, provided by \code{\link{fit_manova}}.
}
\examples{
wafers_folder = file.path(path.package("voltagefit"),"extdata") # path to wafers data directory
fitted = fit_all(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), 
        replicate = 1:6, treatment = rep(1,6))
fitman = fit_manova(fitted, design)
get_params(fitman)

}

