% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histplot.R
\name{histplot}
\alias{histplot}
\title{Plot histograms of fitted wafer curve parameters and cost function}
\usage{
histplot(fitted)
}
\arguments{
\item{fitted}{Fitted data.frame as given by \code{\link{fitwafer}}.}
}
\value{
None
}
\description{
Function to plot histograms of fitted wafer curve parameters and cost function for both forward and backwards curves
}
\examples{
fitted = fitwafer(wafer3737)
histplot(fitted)

}

