% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_functions.R
\name{interp_residuals_derivs}
\alias{interp_residuals}
\alias{interp_residuals_derivs}
\title{Interpolated residuals cost function}
\usage{
interp_residuals_derivs(param, device_model, datax, datay)

interp_residuals(param, datax, datay, device_model)
}
\arguments{
\item{param}{a list consisting of parameters for the model curve.}

\item{device_model}{the model voltage curve function.}

\item{datax}{a range of values over which the voltage curve is evaluated.}

\item{datay}{the true voltage values for the datax values given.}
}
\description{
Cost function to be used during parameter estimation, quantifying the difference 
between a given model and the true data for a certain set of model parameters.
The cost defined by this function is based on a sum of residuals over the datax 
values given Interpolation is used to create evenly spaced points as the input 
datax values may have a varying density of points thoughtout. The first derivatives 
with respect to each parameter can also be calculated.
}

