% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmanova.R
\name{man}
\alias{man}
\title{Perform the MANOVA parts of the analysis}
\usage{
man(data, weight)
}
\arguments{
\item{data}{The data frame of parameter values, appended with week and treatment type for each device}

\item{weight}{The weights to be used in the MANOVA, (applied as 1/weight).}
}
\value{
A list, a MANOVA class item (incorporating weights) and the output of vcov (without weights).
}
\description{
Perform the MANOVA parts of the analysis
}

