% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_manova.R
\name{param_manova}
\alias{param_manova}
\title{Perform MANOVA analysis}
\usage{
param_manova(data, weight)
}
\arguments{
\item{data}{The data frame of parameter values, appended with week and treatment type for each device}

\item{weight}{The weights to be used in the MANOVA, (applied as 1/weight).}
}
\value{
A list, a MANOVA class item (incorporating weights) and the
output of vcov (without weights).
}
\description{
Performs a MANOVA analysis using curves fitted to wafers
}

