% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer_plot.R
\name{plot_curves}
\alias{plot_curves}
\title{Plot fitted voltage curve and wafer curves}
\usage{
plot_curves(x, dir = "Forward", plotit = TRUE)
}
\arguments{
\item{x}{fitted wafer data as from \code{fit_wafer}.}

\item{dir}{Direction of voltage curve.}

\item{plotit}{Default \code{code}.}
}
\description{
Plot fitted voltage curve and the wafer curves the fit was 
generated with.
}

