% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer_plot.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot fitted voltage curve and wafer curves}
\usage{
plot_fit(x, plotit = TRUE)
}
\arguments{
\item{x}{fitted wafer data as from \code{fit_wafer}.}

\item{plotit}{Default \code{TRUE}.}
}
\description{
Plot fitted voltage curve and the wafer curves the fit was 
generated with.
}
