% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot histograms of fitted wafer curve parameters and cost function}
\usage{
plot_hist(fitted)
}
\arguments{
\item{fitted}{Fitted data.frame as given by \code{\link{fit_wafer}}.}
}
\value{
None. This function produces a plot.
}
\description{
Plots several histograms, showing the parameters and cost function resulting
from the fitting of forward and backwards curves to wafers.
}
\examples{
fitted = fit_wafer(wafer3737)
plot_hist(fitted)

}

