% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curves.R
\name{plot_under}
\alias{plot_under}
\title{Plot underlying curves}
\usage{
plot_under(fitted, underc, orig = T, med = F)
}
\arguments{
\item{fitted}{Fitted data.frame as given by \code{\link{fit_wafer}}.}

\item{underc}{Underlying curve as given by \code{\link{calc_curves}}.}

\item{orig}{Boolean value deciding whether the plot should be on the original data scale or the transformed scale (Default: TRUE)}

\item{med}{Boolean value deciding whether the whether the median should be used rather than mean (Default: FALSE)}
}
\value{
None. This function produces a plot.
}
\description{
Plot underlying curves, given by \code{\link{calc_curves}}.
N.B. using the mean with low levels of data can lead to spurious curves, 
the median appears to be more stable!
}
\examples{
wafers_folder = file.path(path.package("voltagefit"),"extdata") # path to wafers data directory
fitted = fit_all(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), replicate = 1:6, treatment = rep(1,6))
fitman = fit_manova(fitted, design)
underp = under_param(fitman)
unders = sample_under_param(underp)
underc = calc_curves(fitted, unders)
plot_under(fitted, underc)

}

