% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curves.R
\name{plot_week}
\alias{plot_week}
\title{Plot curves for each week}
\usage{
plot_week(fitted, wcurves, fm, orig = TRUE)
}
\arguments{
\item{fitted}{Fitted data.frame as given by \code{\link{fit_wafer}}.}

\item{wcurves}{Week curves as given by \code{\link{calc_curves}}.}

\item{fm}{Fitted manova as given by \code{\link{fit_manova}}.}

\item{orig}{Boolean value deciding whether the plot should be on the original data scale or the transformed scale (Default: TRUE)}
}
\value{
None. This function produces a plot.
}
\description{
Plot curves, given by \code{\link{calc_curves}}, for each week provided.
}
\examples{
wafers_folder = file.path(path.package("voltagefit"),"extdata") # path to wafers data directory
fitted = fit_all(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), replicate = 1:6, treatment = rep(1,6))
fitman = fit_manova(fitted, design)
weekp = week_param(fitman)
wcurves = calc_curves(fitted, weekp)
plot_week = function(fitted, wcurves, fitman)

}

