% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotunder}
\alias{plotunder}
\title{Plot underlying curve}
\usage{
plotunder(fitted, underc, orig = T, med = F)
}
\arguments{
\item{fitted}{Fitted data.frame as given by \code{\link{fitwafer}}.}

\item{underc}{Underlying curve as given by \code{\link{curves}}.}

\item{orig}{Boolean value deciding whether the plot should be on the original data scale or the transformed scale (Default: TRUE)}

\item{med}{Boolean value deciding whether the whether the median should be used rather than mean (Default: FALSE)}
}
\value{
None. This function produces a plot.
}
\description{
Plot underlying curve
N.B. using the mean with low levels of data can lead to spurious curves, 
the median appears to be more stable!
}
\examples{
wafers_folder = paste(path.package("voltagefit"),"/extdata/",sep="") # path to wafers data directory
fitted = fitall(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), replicate = c(1:6), treatment = rep(1:6))
fitman = fitmanova(fitted, design)
underp = underparam(fitman)
unders = undercurvesim(underp)
underc = curves(fitted, unders)
plotunder(fitted, underc)

}

