% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotweek}
\alias{plotweek}
\title{Plot curves for each week}
\usage{
plotweek(fitted, wcurves, fm, orig = TRUE)
}
\arguments{
\item{fitted}{Fitted data.frame as given by \code{\link{fitwafer}}.}

\item{wcurves}{Week curves as given by \code{\link{curves}}.}

\item{fm}{Fitted manova as given by \code{\link{fitmanova}}.}

\item{orig}{Boolean value deciding whether the plot should be on the original data scale or the transformed scale (Default: TRUE)}
}
\value{
None. This function produces a plot.
}
\description{
Plot curves for each week
}
\examples{
wafers_folder = paste(path.package("voltagefit"),"/extdata/",sep="") # path to wafters data directory
fitted = fitall(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), replicate = c(1:6), treatment = rep(1:6))
fitman = fitmanova(fitted, design)
weekp = weekparam(fitman)
wcurves = curves(fitted, weekp)
plotweek = function(fitted, wcurves, fitman)

}

