% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Sample underlying curves}
\usage{
sample(x, n, ...)
}
\arguments{
\item{x}{Data.frame as output by \code{\link{fit_manova}}.}

\item{n}{Number of samples required (Default: 100).}

\item{...}{Other arguments (not typically needed).}
}
\value{
A data.frame consisting of the fields:
  \describe{
     \item{direction}{Whether the curve direction is forward or backward}
     \item{X1 ... X6}{The parameters characterising the curve}
     \item{type}{Baseline, treatment, or week.}
  }
}
\description{
Generate sample forward and backward underlying curves.
}
