% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/under_param.R
\name{under_param}
\alias{under_param}
\title{Calculate underlying curve parameters}
\usage{
under_param(fm)
}
\arguments{
\item{fm}{Fitted manova list, as output by \code{\link{fit_manova}}.}
}
\value{
A list with elements \code{param} and \code{var}, where param is a data.frame consisting of the fields:
  \describe{
     \item{direction}{Whether the curve direction is forward or backward}
     \item{X1 ... X6}{The parameters characterising the curve}
  }
  and \code{var} is a list with elements forward and backward, each containing the covariance matrix for use in \code{\link{sample_under_param}}.
}
\description{
Calculates the parameters and covariance matrix that charactarises
both the forward and backward underlying curves, using a fitted manova
provided by \code{\link{fit_manova}}.
}
\examples{
wafers_folder = file.path(path.package("voltagefit"),"extdata") # path to wafers data directory
fitted = fit_all(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), 
      replicate = 1:6, treatment = rep(1,6))
fitman = fit_manova(fitted, design)
underp = under_param(fitman)

}

