% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underparam.R
\name{undercurvesim}
\alias{undercurvesim}
\title{Create sample used for underlying curves}
\usage{
undercurvesim(underp, n = 1000)
}
\arguments{
\item{underp}{Data.frame as output by \code{\link{underparam}}.}

\item{n}{Number of samples required (Default: 1000).}
}
\value{
A data.frame consisting of the fields:
  \describe{
     \item{direction}{Whether the curve direction is forward or backward}
     \item{X1 ... X6}{The parameters characterising the curve}
  }
}
\description{
Create sample used for underlying curves
}
\examples{
wafers_folder = file.path(path.package("voltagefit"),"extdata") # path to wafers data directory
fitted = fitall(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), replicate = 1:6, treatment = rep(1,6))
fitman = fitmanova(fitted, design)
underp = underparam(fitman)
undercurvesim(underp)

}

