% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekparam.R
\name{weekparam}
\alias{weekparam}
\title{Calculate parameters for creating curves for each week}
\usage{
weekparam(fm)
}
\arguments{
\item{fm}{Fitted manova list as output by \code{\link{fitmanova}}.}
}
\value{
A data.frame consisting of the fields:
  \describe{
     \item{direction}{Whether the curve direction is forward or backward}
     \item{week}{The week to which the curve corresponds}
     \item{X1 ... X6}{The parameters characterising the week curve}
  }
}
\description{
Calculate parameters for creating curves for each week
}
\examples{
wafers_folder = file.path(path.package("voltagefit"),"extdata") # path to wafers data directory
fitted = fitall(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), replicate = 1:6, treatment = rep(1,6))
fitman = fitmanova(fitted, design)
weekp = weekparam(fitman)

}

