% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wafer.R
\name{fit_all}
\alias{fit_all}
\title{Fit logcurves to multiple wafers}
\usage{
fit_all(path, iterlim = 2000, verbose = TRUE)
}
\arguments{
\item{path}{Directory containing multiple .rds files, each containing a single wafer}

\item{iterlim}{Maxiumum number of iterations to use in nlm (Default: 2000).}
}
\value{
A data.frame consisting of the fields:
  \describe{
     \item{id}{Wafer ID}
     \item{name}{Each device on the wafer}
     \item{max}{Maximum of \code{log(abs(ID))}, where \code{ID} is the current drain data taken from \code{wafer}}
     \item{cost}{The value of the cost function for each device}
     \item{direction}{Whether the curve direction is forward or backward}
     \item{X1 ... X6}{The parameters characterising the curve}
  }
  The attribute v (voltage gate readings for one device), used in curve
  functions and plotting functions, is also appended.
}
\description{
Fits logcurves to multiple wafers in a directory, and returns the parameters
and cost for both forward and backwards curves. This is an extended version
of \code{\link{fit_wafer}}. It is assumed that any wafer entering this
function (and subsequently fit_wafer) has been pre-filtered. See the function
"filtered" for details of how wafers are filtered.
}
\examples{
wafers_folder = file.path(path.package("voltagefit"),"extdata") # path to wafers data directory
fit = fit_all(wafers_folder)

}

