% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wafer.R
\name{fit_wafer}
\alias{fit_wafer}
\title{Fit logcurves to a wafer}
\usage{
fit_wafer(wafer, maxit = 10000, verbose = TRUE)
}
\arguments{
\item{wafer}{Data from a single wafer.}

\item{iterlim}{Maxiumum number of iterations to use in nlm (Default: 2000).}

\item{id}{WaferID (Default: A random string is generated as the ID).}
}
\value{
A data.frame consisting of the fields:
  \describe{
     \item{id}{Wafer ID}
     \item{name}{Each device on the wafer}
     \item{max}{Maximum of \code{log(abs(ID))}, where \code{ID} is the current drain data taken from \code{wafer}}
     \item{cost}{The value of the cost function for each device}
     \item{direction}{Whether the curve direction is forward or backward}
     \item{X1 ... X6}{The parameters characterising the curve}
  }
}
\description{
Fits logcurves to a wafer and returns the curve parameters and resulting
cost for both forward and backwards curves. It is assumed that any wafer
entering this function has been pre-filtered. See the function "filtered"
for details of how wafers are filtered.
}
\examples{
fit_wafer(wafer3737)

}

