% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer_fit.R
\name{trans_device}
\alias{fit_wafer}
\alias{trans_device}
\alias{validate_device}
\title{Fit curves to a wafer}
\usage{
trans_device(wafer)

validate_device(wafer)

fit_wafer(wafer, trans = trans_device, validate = validate_device,
  cost_func = area_between_curves, dev_curve = curve_4BARO,
  initparams = c(-10, -30, 5, 0), maxit = 10000, verbose = TRUE,
  plot = FALSE)
}
\arguments{
\item{wafer}{Data from a single wafer}

\item{trans}{(Optional) Wafer ransform function}

\item{validate}{(Optional) Wafer validate function}

\item{cost_func}{(Optional) Cost function}

\item{dev_curve}{(Optional) Model device voltage curve}

\item{initparams}{(Optional) Inital parameter estimation}

\item{maxit}{(Optional) Maxiumum number of iterations for use in optimiser (Default: 10000)}

\item{verbose}{(Optional) Print verbose output (Default: \code{TRUE})}

\item{plot}{(Optional) Plot devices with fitted model (Default: \code{FALSE})}
}
\value{
A data.frame consisting of the fields:
  \describe{
     \item{id}{Wafer ID}
     \item{name}{Each device on the wafer}
     \item{cost}{The value of the cost function for each device}
     \item{direction}{Whether the curve direction is forward or backward}
     \item{X1 ... X6}{The parameters characterising the curve}
  }
}
\description{
Fits curves to a wafer and returns the curve parameters and resulting
cost for both forward and backwards curves. Wafers can be validated and
transformed before fitting.
}
\examples{
fit_wafer(wafer3737)

}

