% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer_fit.R
\name{trans_device}
\alias{trans_device}
\alias{validate_device}
\alias{trans_filter_device}
\alias{fit_wafer}
\title{Fit curves to a wafer}
\usage{
trans_device(wafer)

validate_device(wafer)

trans_filter_device(wafer, validate = validate_device, trans = trans_device,
  verbose = TRUE)

fit_wafer(wafer, trans = trans_device, validate = validate_device,
  cost_func = area_between_curves, dev_curve = curve_piecewise,
  initparams = NULL, maxit = 1e+05, verbose = TRUE, plot = FALSE)
}
\arguments{
\item{wafer}{Data from a single wafer}

\item{validate}{(Optional) Wafer validate function}

\item{trans}{(Optional) Wafer ransform function}

\item{verbose}{(Optional) Print verbose output (Default: \code{TRUE})}

\item{cost_func}{(Optional) Cost function}

\item{dev_curve}{(Optional) Model device voltage curve}

\item{initparams}{(Optional) Inital parameter estimation}

\item{maxit}{(Optional) Maxiumum number of iterations for use in optimiser (Default: 10000)}

\item{plot}{(Optional) Plot all devices and fitted curves (Default: \code{FALSE})}
}
\value{
A data.frame consisting of the fields:
  \describe{
     \item{id}{Wafer ID}
     \item{name}{Each device on the wafer}
     \item{cost}{The value of the cost function for each device}
     \item{direction}{Whether the curve direction is forward or backward}
     \item{X1 ... X6}{The parameters characterising the curve}
  }
}
\description{
Fits curves to a wafer and returns the curve parameters and resulting
cost for both forward and backwards curves. Wafers can be validated and
transformed before fitting. Devices are fitted in parallel automatically.
}
\examples{
data(voltage, package="voltagefit")
\dontrun{
fit_wafer(wafer6138)
}

}
