% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fitall}
\alias{fitall}
\title{Fit logcurve to multiple wafers}
\usage{
fitall(path, iterlim = 2000)
}
\arguments{
\item{path}{Directory where multiple .rds files, each containing a single wafer, are located}

\item{iterlim}{Maxiumum number of iterations to use in nlm (Default: 2000).}
}
\value{
A data.frame consisting of the fields:
  \describe{
     \item{id}{Wafer ID}
     \item{name}{Each device on the wafer}
     \item{max}{Maximum of \code{log(abs(ID))}, where \code{ID} is the current drain data taken from \code{wafer}}
     \item{cost}{The value of the cost function for each device}
     \item{direction}{Whether the curve direction is forward or backward}
     \item{X1 ... X6}{The parameters characterising the curve}
  }
  The attribute v (voltage gate readings for one device), used in curve functions and plotting functions, is also appended.
}
\description{
Fits logcurves to multiple wafers, given in a directory, and returns the parameters and cost 
for both forward and backwards curves. An extended version of fitwafer. It is assumed that 
any wafer entering this function (and subsequently fitwafer) has been pre-filtered. 
See the function "filtered" for details of how wafers are filtered.
}
\examples{
wafers_folder = file.path(path.package("voltagefit"),"extdata") # path to wafers data directory
fit = fitall(wafers_folder)

}

