% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmanova.R
\name{fitmanova}
\alias{fitmanova}
\title{Fit the results of fitwafer using MANOVA}
\usage{
fitmanova(fitted, design)
}
\arguments{
\item{fitted}{Fitted data.frame as given by \code{\link{fitwafer}}.}

\item{design}{The design matrix (of type data.frame with columns week, wafer, replicate, treatment).}
}
\value{
A list with elements forward and backward, each containing the MANOVA output incorporating 
weights and vcov (excluding weights).
}
\description{
Fit the results of fitwafer using MANOVA
}
\examples{
wafers_folder = paste(path.package("voltagefit"),"/extdata/",sep="") # path to wafters data directory
fitted = fitall(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), replicate = c(1:6), treatment = rep(1:6))
fitman = fitmanova(fitted, design)

}

