% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_functions.R
\name{cost_abc_derivs}
\alias{area_between_curves}
\alias{cost_abc_derivs}
\alias{interp_residuals}
\title{Cost functions}
\usage{
cost_abc_derivs(param, device_model, datax, datay)

interp_residuals(param, datax, datay, device_model)

area_between_curves(param, datax, datay, device_model)
}
\arguments{
\item{param}{list of parameters for the model voltage curve function}

\item{device_model}{the model voltage curve function}

\item{datax}{x values resulting from a measurement of a device}

\item{datay}{y values resulting from a measurement of a device}
}
\value{
A value representing the "cost" associated with the given model and parameters.
}
\description{
Cost functions are used during parameter estimation, and in general quantifiy
the difference between a given model and true data, for a certain set of model parameters.
}

