% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Sample underlying curves}
\usage{
sample(x, n, ...)
}
\arguments{
\item{x}{Data.frame as output by \code{\link{fit_manova}}.}

\item{n}{Number of samples required (Default: 100).}

\item{...}{Other arguments (not typically needed).}
}
\value{
A data.frame consisting of the fields:
  \describe{
     \item{direction}{Whether the curve direction is forward or backward}
     \item{X1 ... X6}{The parameters characterising the curve}
     \item{type}{Baseline, treatment, or week.}
  }
}
\description{
Generate sample forward and backward underlying curves.
}
\examples{
wafers_folder = file.path(path.package("voltagefit"),"extdata") # path to wafers data directory
fitted = fit_all(wafers_folder,dev_curve=curve_5BARO)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), 
        replicate = 1:6, treatment = rep(1,6))
fitman = fit_manova(fitted, design)
samples = sample(fitman)
}

