% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underparam.R
\name{undercurvesim}
\alias{undercurvesim}
\title{Create sample used for underlying curves}
\usage{
undercurvesim(underp, n = 1000)
}
\arguments{
\item{underp}{Data.frame as output by \code{\link{underparam}}.}

\item{n}{Number of samples required (default=1000).}
}
\value{
A data.frame consisting of the fields:\cr
  direction - Forward or backward curve.\cr
  X1 ... X6 - The parameters characterising the curves.
}
\description{
Create sample used for underlying curves
}
\examples{
wafers_folder = paste(path.package("voltagefit"),"/extdata/",sep="") # path to wafers data directory
fitted = fitall(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), replicate = c(1:6), treatment = rep(1:6))
fitman = fitmanova(fitted, design)
underp = underparam(fitman)
undercurvesim(underp)

}

