% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underparam.R
\name{underparam}
\alias{underparam}
\title{Calculate parameters for creating the underlying curve}
\usage{
underparam(fm)
}
\arguments{
\item{fm}{Fitted manova list as output by \code{\link{fitmanova}}.}
}
\value{
A list with elements param and var.\cr\cr
  param - A data.frame consisting of the fields:\cr
  direction - Forward or backward curve.\cr
  X1 ... X6 - The parameters characterising the week curves.\cr\cr
  var - A list with elements forward and backward, each containing the covariance matrix for use in \code{\link{undercurvesim}}.
}
\description{
Calculate parameters for creating the underlying curve
}
\examples{
wafers_folder = paste(path.package("voltagefit"),"/extdata/",sep="") # path to wafers data directory
fitted = fitall(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), replicate = c(1:6), treatment = rep(1:6))
fitman = fitmanova(fitted, design)
underp = underparam(fitman)

}

