% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekparam.R
\name{weekparam}
\alias{weekparam}
\title{Calculate parameters for creating curves for each week}
\usage{
weekparam(fm)
}
\arguments{
\item{fm}{Fitted manova list as output by \code{\link{fitmanova}}.}
}
\value{
A data.frame consisting of the fields:\cr
  direction - Forward or backward curve.\cr
  week  - The week to which the curve corresponds.\cr
  X1 ... X6 = The parameters characterising the week curves.\cr
}
\description{
Calculate parameters for creating curves for each week
}
\examples{
wafers_folder = paste(path.package("voltagefit"),"/extdata/",sep="") # path to wafters data directory
fitted = fitall(wafers_folder)
design = data.frame(week = c(1,1,1,1,2,2), wafer = unique(fitted$id), replicate = c(1:6), treatment = rep(1:6))
fitman = fitmanova(fitted, design)
weekp = weekparam(fitman)

}

